pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-triage-env"
version = "1.0.0"
description = "A deterministic MERN stack triage simulator for SRE agents"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic",
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
    "httpx",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = ["pytest"]
